horarios discriminan cargos. mejora gui. correccion error en switch carg

+++ cargos/models.py
@@ -1,4 +1,5 @@
 from django.db import models
+#from login.models import CustomUser
 
 # Create your models here.
 
@@ -9,6 +10,7 @@ class CargosCache(models.Model):
     desc_regional = models.CharField(max_length=200)
     desc_categ = models.CharField(max_length=200)
     desc_dedic = models.CharField(max_length=200)
-    horas_dedicacion = models.IntegerField()
+    horas_dedicacion = models.FloatField()
     escalafon = models.CharField(max_length=200) 
     seleccionado = models.BooleanField(default=False)
+    #agente = models.ManyToManyField(CustomUser)
